+++ VoiceInk/Services/UserDefaultsManager.swift
@@ -5,8 +5,6 @@ extension UserDefaults {
         static let audioInputMode = "audioInputMode"
         static let selectedAudioDeviceUID = "selectedAudioDeviceUID"
         static let selectedAudioDeviceModelUID = "selectedAudioDeviceModelUID"
-        static let prioritizedDevices = "prioritizedDevices"
-        static let affiliatePromotionDismissed = "VoiceInkAffiliatePromotionDismissed"
     }
 
     var audioInputModeRawValue: String? {
@@ -24,13 +22,4 @@ extension UserDefaults {
         set { setValue(newValue, forKey: Keys.selectedAudioDeviceModelUID) }
     }
 
-    var prioritizedDevicesData: Data? {
-        get { data(forKey: Keys.prioritizedDevices) }
-        set { setValue(newValue, forKey: Keys.prioritizedDevices) }
-    }
-
-    var affiliatePromotionDismissed: Bool {
-        get { bool(forKey: Keys.affiliatePromotionDismissed) }
-        set { setValue(newValue, forKey: Keys.affiliatePromotionDismissed) }
-    }
 }