+++ VoiceInk/Transcription/Engine/RecordingState.swift
@@ -5,6 +5,4 @@ enum RecordingState: Equatable {
     case starting
     case recording
     case transcribing
-    case enhancing
-    case busy
 }